+++ multisrc/overrides/madara/manhuabox/src/ManhuaBox.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.manhuabox
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+
+class ManhuaBox : Madara("ManhuaBox", "https://manhuabox.net", "en") {
+    override val pageListParseSelector = "div.page-break, div.text-left p img"
+}
