+++ multisrc/overrides/madara/manhuaplus/src/ManhuaPlus.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.manhuaplus
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+
+class ManhuaPlus : Madara("Manhua Plus", "https://manhuaplus.com", "en") {
+    override val pageListParseSelector = "li.blocks-gallery-item"
+}
