+++ multisrc/overrides/madara/manhuaus/src/ManhuaUS.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.en.manhuaus
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+
+class ManhuaUS : Madara("ManhuaUS", "https://manhuaus.com", "en") {
+    override val pageListParseSelector = "li.blocks-gallery-item"
+}
