+++ multisrc/overrides/madara/martialscans/src/MartialScans.kt
@@ -1,0 +1,26 @@
+package eu.kanade.tachiyomi.extension.en.martialscans
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import eu.kanade.tachiyomi.source.model.SManga
+import org.jsoup.nodes.Element
+
+class MartialScans : Madara("Martial Scans", "https://martialscans.com", "en") {
+    override fun popularMangaFromElement(element: Element): SManga {
+        val manga = SManga.create()
+
+        with(element) {
+            select(popularMangaUrlSelector).last()?.let {
+                manga.setUrlWithoutDomain(it.attr("href"))
+                manga.title = it.ownText()
+            }
+
+            select("img").last()?.let {
+                manga.thumbnail_url = imageFromElement(it)
+            }
+        }
+
+        return manga
+    }
+
+    override fun searchMangaFromElement(element: Element): SManga = popularMangaFromElement(element)
+}
