+++ multisrc/overrides/madara/mundowuxia/src/MundoWuxia.kt
@@ -1,0 +1,7 @@
+package eu.kanade.tachiyomi.extension.es.mundowuxia
+
+import eu.kanade.tachiyomi.multisrc.madara.Madara
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+class MundoWuxia : Madara("Mundo Wuxia", "https://mundowuxia.com", "es", SimpleDateFormat("MMMM dd, yyyy", Locale("es")))
