+++ multisrc/src/main/java/eu/kanade/tachiyomi/multisrc/madara/MadaraGenerator.kt
@@ -80,7 +80,30 @@ class MadaraGenerator : ThemeSourceGenerator {
             SingleLang("Kombatch", "https://kombatch.com", "id"),
 
             SingleLang("Mangazuki.online", "http://mangazukinew.online", "en", className = "MangazukiOnline"),
-
+            SingleLang("ManhuaBox", "https://manhuabox.net", "en"),
+            SingleLang("ManhuaFast", "https://manhuafast.com", "en"),
+            SingleLang("Manhuaga", "https://manhuaga.com", "en"),
+            SingleLang("Manhua Plus", "https://manhuaplus.com", "en"),
+            SingleLang("Manhuas.net", "https://manhuas.net", "en"),
+            SingleLang("Manhuas World", "https://manhuasworld.com", "en"),
+            SingleLang("Manhua SY", "https://www.manhuasy.com", "en"),
+            SingleLang("ManhuaUS", "https://manhuaus.com", "en"),
+            SingleLang("Manhwa Raw", "https://manhwaraw.com", "ko"),
+            SingleLang("Manhwatop", "https://manhwatop.com", "en"),
+            SingleLang("Manwahentai.me", "https://manhwahentai.me", "en"),
+            SingleLang("Manwha Club", "https://manhwa.club", "en"),
+            SingleLang("ManyToon", "https://manytoon.com", "en"),
+            SingleLang("ManyToonClub", "https://manytoon.club", "ko"),
+            SingleLang("ManyToon.me", "https://manytoon.me", "en"),
+            SingleLang("Mark Scans", "https://markscans.online", "pt-BR"),
+            SingleLang("Martial Scans", "https://martialscans.com", "en"),
+            SingleLang("MG Komik", "https://mgkomik.my.id", "id"),
+            SingleLang("Milftoon", "https://milftoon.xxx", "en"),
+            SingleLang("Miracle Scans", "https://miraclescans.com", "en"),
+            SingleLang("Mixed Manga", "https://mixedmanga.com", "en"),
+            SingleLang("MMScans", "https://mm-scans.com/", "en"),
+            SingleLang("Mundo Wuxia", "https://mundowuxia.com", "es"),
+            SingleLang("Mystical Merries", "https://mysticalmerries.com", "en"),
             SingleLang("Nazarick Scans", "https://nazarickscans.com", "en"),
             SingleLang("NeatManga", "https://neatmanga.com", "en"),
             SingleLang("NekoBreaker", "https://nekobreaker.com", "pt-BR"),
